Extract consts into folder
