+++ src/utils/consts/icons.ts
@@ -1,0 +1,28 @@
+export type IconName =
+  | 'AddNew'
+  | 'Remove'
+  | 'RemoveFAB'
+  | 'Waste'
+  | 'WasteFAB'
+  | 'MoveToShoppingList'
+  | 'Purchase'
+  | 'TakePhoto'
+  | 'Statistics'
+  | 'Storage'
+  | 'Recipes'
+  | 'ShoppingList';
+
+export const Icons: { [iconName in IconName]: string } = {
+  AddNew: require('@/assets/images/AddNew.svg'),
+  Remove: require('@/assets/images/Cross.svg'),
+  RemoveFAB: require('@/assets/images/RemoveFAB.svg'),
+  Waste: require('@/assets/images/Waste.svg'),
+  WasteFAB: require('@/assets/images/WasteFAB.svg'),
+  MoveToShoppingList: require('@/assets/images/MoveToShoppingListFAB.svg'),
+  Purchase: require('@/assets/images/PurchaseFAB.svg'),
+  TakePhoto: require('@/assets/images/TakePhoto.svg'),
+  Statistics: require('@/assets/images/Statistics.svg'),
+  Storage: require('@/assets/images/Storage.svg'),
+  ShoppingList: require('@/assets/images/ShoppingList.svg'),
+  Recipes: require('@/assets/images/Recipes.svg')
+};
